Move things to a separate util functions file
